+++ packages/pluggableWidgets/rich-text-web/typings/RichTextProps.d.ts
@@ -23,12 +23,22 @@ export type OverflowYEnum = "auto" | "scroll" | "hidden";
 
 export type OnChangeTypeEnum = "onLeave" | "onDataChange";
 
+export interface CustomFontsType {
+    fontName: string;
+    fontStyle: string;
+}
+
 export type ToolbarConfigEnum = "basic" | "advanced";
 
 export type CtItemTypeEnum = "separator" | "undo" | "redo" | "bold" | "italic" | "underline" | "strike" | "superScript" | "subScript" | "orderedList" | "bulletList" | "lowerAlphaList" | "checkList" | "minIndent" | "plusIndent" | "direction" | "link" | "image" | "video" | "formula" | "blockquote" | "code" | "codeBlock" | "viewCode" | "align" | "centerAlign" | "rightAlign" | "font" | "size" | "color" | "background" | "header" | "fullscreen" | "clean" | "tableBetter";
 
 export interface AdvancedConfigType {
     ctItemType: CtItemTypeEnum;
+}
+
+export interface CustomFontsPreviewType {
+    fontName: string;
+    fontStyle: string;
 }
 
 export interface AdvancedConfigPreviewType {
@@ -59,6 +69,7 @@ export interface RichTextContainerProps {
     onLoad?: ActionValue;
     onChangeType: OnChangeTypeEnum;
     spellCheck: boolean;
+    customFonts: CustomFontsType[];
     toolbarConfig: ToolbarConfigEnum;
     history: boolean;
     fontStyle: boolean;
@@ -100,6 +111,7 @@ export interface RichTextPreviewProps {
     onLoad: {} | null;
     onChangeType: OnChangeTypeEnum;
     spellCheck: boolean;
+    customFonts: CustomFontsPreviewType[];
     toolbarConfig: ToolbarConfigEnum;
     history: boolean;
     fontStyle: boolean;